+++ library/Gc/Mvc/Module.php
@@ -27,8 +27,6 @@
 
 namespace Gc\Mvc;
 
-use Gc\Core\Config as CoreConfig;
-
 /**
  * Generic Module
  *