+++ tests/config/prepare-database.php
@@ -24,8 +24,6 @@
  * @link     http://www.got-cms.com
  */
 
-use Gc\Core\Config as CoreConfig;
-use Gc\Registry;
 use Zend\Db\Adapter\Adapter as DbAdapter;
 use Zend\Db\TableGateway\TableGateway;
 
@@ -48,7 +46,6 @@ try {
 }
 
 \Zend\Db\TableGateway\Feature\GlobalAdapterFeature::setStaticAdapter($dbAdapter);
-CoreConfig::getInstance();
 
 /**
  * Install database