+++ tests/library/Gc/View/Helper/CdnTest.php
@@ -26,7 +26,6 @@
 
 namespace Gc\View\Helper;
 
-use Gc\Core\Config as CoreConfig;
 use Zend\Form\Element;
 use Gc\Registry;
 
@@ -52,7 +51,10 @@ class CdnTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-        $this->object = new Cdn(Registry::get('Application')->getRequest());
+        $this->object = new Cdn(
+            Registry::get('Application')->getRequest(),
+            Registry::get('Application')->getServiceManager()->get('CoreConfig')
+        );
     }
 
     /**
@@ -73,8 +75,9 @@ class CdnTest extends \PHPUnit_Framework_TestCase
      */
     public function testInvokeSecure()
     {
-        $basePath = CoreConfig::setValue('force_frontend_ssl', 1);
-        $basePath = CoreConfig::setValue('secure_cdn_base_path', 'https://got-cms.com');
+        $coreConfig = Registry::get('Application')->getServiceManager()->get('CoreConfig');
+        $basePath = $coreConfig->setValue('force_frontend_ssl', 1);
+        $basePath = $coreConfig->setValue('secure_cdn_base_path', 'https://got-cms.com');
         $this->assertEquals('https://got-cms.com/test', $this->object->__invoke('test'));
     }
 
@@ -85,8 +88,9 @@ class CdnTest extends \PHPUnit_Framework_TestCase
      */
     public function testInvokeUnsecure()
     {
-        $basePath = CoreConfig::setValue('force_frontend_ssl', 0);
-        $basePath = CoreConfig::setValue('unsecure_cdn_base_path', 'http://got-cms.com');
+        $coreConfig = Registry::get('Application')->getServiceManager()->get('CoreConfig');
+        $basePath = $coreConfig->setValue('force_frontend_ssl', 0);
+        $basePath = $coreConfig->setValue('unsecure_cdn_base_path', 'http://got-cms.com');
         $this->assertEquals('http://got-cms.com/test', $this->object->__invoke('test'));
     }
 }